サインアップページと、そのPOSTメソッドでデータベースにユーザ情報を保存する機能を実装（セキュリティ度外視）

+++ server.js
@@ -1,7 +1,10 @@
 const http = require('http');
 const { isDeepStrictEqual } = require('util');
-const { topPage, postPage, notFoundPage } = require('./pages');
+const { signUpPage, postSignUpPage, signInPage, topPage, getPostPage, postPostPage, notFoundPage } = require('./pages');
 const { deletePost, db } = require('./databaseUtils');
+
+//セッション
+const sessions = {};
 
 const hostname = '127.0.0.1';
 const PORT = 3000;
@@ -16,29 +19,55 @@ const server = http.createServer((req, res) => {
   //ルーティング
   const id = req.url.split('/').pop(); //URLから削除対象のIDを取得
 
-  switch (req.url) { //リクエストされたurlが引数に入る
-    case '/':
-      topPage(req, res); //トップページ用の関数を呼んでいる
-      break;
-    case '/post':
-      postPage(req, res); //投稿用ページの関数を呼んでいる
-      break;
-    case `/posts/${id}`: //削除を実行
-      deletePost(id, (err) => {
-        if (err) {
-          console.error(err.message);
-          res.statusCode = 500;
-          res.end('削除時にエラーが発生しました');
-          return;
-        }
-        console.log('投稿を削除しました');
-        res.statusCode = 200;
-        res.end('削除が完了しました');
-      });
-      break;
-    default:
-      notFoundPage(req, res); //その他のリクエストをNot Foundページとして表示
-      break
+  if (req.method === 'GET') {
+    switch (req.url) { //リクエストされたurlが引数に入る
+      case '/':
+        topPage(req, res); //トップページ用の関数を呼んでいる
+        break;
+      case '/sign_up':
+        signUpPage(req, res);
+        break
+      case '/sign_in':
+        signInPage(req, res);
+        break
+      case '/post':
+        getPostPage(req, res); //投稿用ページの関数を呼んでいる
+        break;
+      default:
+        notFoundPage(req, res); //その他のリクエストをNot Foundページとして表示
+        break;
+    }
+  } else if (req.method === 'POST') {
+    switch (req.url) {
+      case '/post':
+        postPostPage(req, res);
+        break;
+      case '/sign_up':
+        postSignUpPage(req, res);
+        break;
+      default:
+        notFoundPage(req, res);
+        break;
+    }
+  } else if (req.method === 'DELETE') {
+    switch (req.url) {
+      case `/posts/${id}`: //削除を実行
+        deletePost(id, (err) => {
+          if (err) {
+            console.error(err.message);
+            res.statusCode = 500;
+            res.end('削除時にエラーが発生しました');
+            return;
+          }
+          console.log('投稿を削除しました');
+          res.statusCode = 200;
+          res.end('削除が完了しました');
+        });
+        break;
+      default:
+        notFoundPage(req, res); //その他のリクエストをNot Foundページとして表示
+        break;
+    }
   }
 });
 
@@ -47,6 +76,7 @@ server.listen(PORT, hostname, () => {
   console.log(`Server running at http://${hostname}:${PORT}`);
 });
 
+// ctrl + C でサーバシャットダウン時にdbを閉じる
 process.on('SIGINT', () => {
   db.close((err) => {
     if (err) {
